refactor(filters): split filters.py into focused package (Phase 2)
Convert advanced_alchemy/filters.py (1508 lines) into a six-submodule
package with re-export parity. No behavioral change; every existing
public name still imports from advanced_alchemy.filters.

Layout:
  filters/_base.py          ABCs, type vars, FilterMap/LogicalOperatorMap
  filters/_pagination.py    LimitOffset, OrderBy
  filters/_search.py        SearchFilter, NotInSearchFilter
  filters/_relationship.py  RelationshipFilter
  filters/_columns.py       BeforeAfter, OnBeforeAfter, Collection*, Null*,
                            ComparisonFilter, operators_map, VALID_OPERATORS
  filters/_logical.py       ExistsFilter, NotExistsFilter, FilterGroup,
                            MultiFilter
  filters/__init__.py       Re-exports + FilterTypes TypeAlias

CollectionFilter relies on a lazy import of RelationshipFilter to avoid a
circular dep — the relationship-handling path is only reached when the
field is a relationship attribute.

Tests:
- New tests/unit/test_filters/test_public_surface.py: 74 parametrized
  cases asserting every name in __all__ imports from the public package
  and resolves to the same object as imported via its expected submodule.
- tests/integration/test_filters.py adjusted: import and_/or_ from
  sqlalchemy directly rather than relying on the old single-file
  filters.py to leak them through.

Verification: mypy clean, pyright clean, ruff clean, 1388 unit tests +
41 sqlite filter integration tests pass.

Refs: relationship-filters PRD v2.0 Phase 2 (advanced-alchemy-38c.10)
