+++ advanced_alchemy/filters/_pagination.py
@@ -1,0 +1,113 @@
+"""Pagination and ordering filters."""
+
+from dataclasses import dataclass
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    Literal,
+    Union,
+    cast,
+)
+
+from sqlalchemy import ColumnElement, Select
+
+from advanced_alchemy.filters._base import (
+    ModelT,
+    PaginationFilter,
+    StatementFilter,
+    StatementTypeT,
+)
+
+if TYPE_CHECKING:
+    from sqlalchemy.orm import InstrumentedAttribute
+
+__all__ = ("LimitOffset", "OrderBy")
+
+
+@dataclass
+class LimitOffset(PaginationFilter):
+    """Limit and offset pagination filter.
+
+    Implements traditional pagination using SQL LIMIT and OFFSET clauses.
+    Only applies to SELECT statements; other statement types are returned unmodified.
+
+    Note:
+        This filter only modifies SELECT statements. For other statement types
+        (UPDATE, DELETE), the statement is returned unchanged.
+
+    See Also:
+        - :meth:`sqlalchemy.sql.expression.Select.limit`: SQLAlchemy LIMIT clause
+        - :meth:`sqlalchemy.sql.expression.Select.offset`: SQLAlchemy OFFSET clause
+    """
+
+    limit: int
+    """Maximum number of rows to return."""
+    offset: int
+    """Number of rows to skip before returning results."""
+
+    def append_to_statement(self, statement: StatementTypeT, model: type[ModelT]) -> StatementTypeT:
+        """Apply LIMIT/OFFSET pagination to the statement.
+
+        Args:
+            statement: The SQLAlchemy statement to modify
+            model: The SQLAlchemy model class
+
+        Returns:
+            StatementTypeT: Modified statement with limit and offset applied
+
+        Note:
+            Only modifies SELECT statements. Other statement types are returned as-is.
+
+        See Also:
+            :class:`sqlalchemy.sql.expression.Select`: SQLAlchemy SELECT statement
+        """
+        if isinstance(statement, Select):
+            statement = cast("StatementTypeT", statement.limit(self.limit).offset(self.offset))
+        return statement
+
+
+@dataclass
+class OrderBy(StatementFilter):
+    """Order by a specific field.
+
+    Appends an ORDER BY clause to SELECT statements, sorting records by the
+    specified field in ascending or descending order.
+
+    Note:
+        This filter only modifies SELECT statements. For other statement types,
+        the statement is returned unchanged.
+
+    See Also:
+        - :meth:`sqlalchemy.sql.expression.Select.order_by`: SQLAlchemy ORDER BY clause
+        - :meth:`sqlalchemy.sql.expression.ColumnElement.asc`: Ascending order
+        - :meth:`sqlalchemy.sql.expression.ColumnElement.desc`: Descending order
+    """
+
+    field_name: "Union[str, ColumnElement[Any], InstrumentedAttribute[Any]]"
+    """Field name, model attribute, or func expression (e.g., ``func.random()``)."""
+    sort_order: Literal["asc", "desc"] = "asc"
+    """Sort direction ("asc" or "desc")."""
+
+    def append_to_statement(self, statement: StatementTypeT, model: type[ModelT]) -> StatementTypeT:
+        """Append an ORDER BY clause to the statement.
+
+        Args:
+            statement: The SQLAlchemy statement to modify
+            model: The SQLAlchemy model class
+
+        Returns:
+            StatementTypeT: Modified statement with an ORDER BY clause
+
+        Note:
+            Only modifies SELECT statements. Other statement types are returned as-is.
+
+        See Also:
+            :meth:`sqlalchemy.sql.expression.Select.order_by`: SQLAlchemy ORDER BY
+        """
+        if isinstance(statement, Select):
+            field = self._get_instrumented_attr(model, self.field_name)
+            if self.sort_order == "desc":
+                statement = cast("StatementTypeT", statement.order_by(field.desc()))
+            else:
+                statement = cast("StatementTypeT", statement.order_by(field.asc()))
+        return statement
